Implemented n-way merge strategy.
It's a bit fragile ;-)

It works only for files, not for directories. Conditions are not
provided by the user right now but simply generated in a very broken and
ugly way.

Also, it duplicates code instead of extending conditions, which is,
again, ugly. But no time right now.

All in all it's a real quick hack and needs a lot of fixing and
polishing, but it gives an idea on how n-way merging can be done.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>