Fix pretty-printing of conflicts.
This commit fixes several issues related to pretty-printing
of conflicts. There were cases were a conflict was detected on
the AST-level, but not printed by the pretty-printer. Affected
were several language elements that had been introduced by
Java 5 or newer.

Only versions using the fairly new aspect-oriented extension of
the JastAddJ pretty-printer are affected.
Versions of JDime <0.3.8 are not affected.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>