Removed Artifact.createArtifact(boolean isLeaf).
This method was actually only implemented by FileArtifact. Calls to this
function only intended to make sure the outputfile of the merge (target)
is created if necessary.

Now this is handled in the constructors of FileArtifact, so the method
from the super class Artifact as well as the respective calls have been
removed.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>