Add an unordered matcher using the hungarian method.
This matcher uses the hungarian algorithm to solve the assignment
problem in O(n^3). It can be used instead of the LPMatcher, which uses
glpk via jni.

The code for the actual algorithm is provided by Kevin Stern under the
MIT license. It is currently hosted on GitHub:
https://github.com/KevinStern/software-and-algorithms

Pleas note that this commit introduces several duplicate code with
LPMatcher, so a refactoring is due soon.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>