Fix issue 7
The issue in question occurs in a part of the AST that is not relevant
for the merge result, i.e. this part is not pretty-printed anyway.

In the case of the respective issue, a runtime check fails in
rebuildAST() because a deletion-deletion situation ends in incorrect
syntax (second part of a dot element missing). If we ignore this
failure, flushCaches() will also fail because of the missing element.
However, during prettyprinting the whole subtree is ignored whatsoever,
so it is save to ignore all those failures at least in this case.

This patch contains a hack that catches any exceptions from rebuildAST()
and flushCaches(), and instead uses a very simple heuristic to check
whether pretty-printing failed (i.e., it checks whether the output is
empty) or not.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>