Fix OrderedMerge: first remaining child was missed
After the main loop, we handle remaining children on one of the
competing sides (i.e., either the left or right site is empty, the other
one has unmerged children remaining).

However, we missed the first of these children, as we immediately
executed .next() at the beginning of the loop.

This patch gets the loop right.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>