Change behaviour if output files already exist
So far, we deleted existing output files *before* executing the merge,
if the parameter --force-overwrite (-f) was specified.
This leads to problems with inplace merges, where input files
are specified to be replaced by the merge output, because at
merge-time the input files have been already deleted.

This patch removes the upfront deletion, as well as an existing check
in FileArtifact. If --force-overwrite is specified, files are now
overwritten *after* the merge.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>