Fix missing cases in Merge.java
Merge.java usually delegates the merge to either OrderedMerge or
UnorderedMerge, depending on the type of the merged element.
However, in some "simple" cases, the delegation is not performed,
but Merge.java handles the situation itself.
This happens if one of the competing elements has no children.

In this opimization, we missed cases where a necessary deletion
was not performed.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>