Only construct MergeScenarioStatistics in Statistics
This ensures that they are always registered in a Statistics instance
and there are no unnecessary calls to the add function which might
cause some fields of the MergeScenarioStatistics to be lost.

It also fixes some issues that turned up while adding a statistics
field for recording the name of the MergeStrategy that handled a given
MergeScenario.